refactored and reworked. everything works how it is supposed to but still no local storage
