+++ src/NewProject.js
@@ -1,0 +1,10 @@
+
+
+export default class Project {
+    constructor(name, id){
+        this.name = name
+        this.id = id
+        this.arr = []
+    }
+}
+
