+++ src/NewTodo.js
@@ -1,0 +1,15 @@
+
+
+
+
+export default class Todo {
+    constructor(title, date, priority, projectId, id) {
+        this.title = title
+        this.date = date
+        this.priority = priority
+        this.projectId = projectId
+        this.id = id
+    }
+
+
+}
