+++ src/renderProjectTodos.js
@@ -1,5 +1,6 @@
 import { projects } from "./index"
-import { defaultProjectArr } from "./index"
+import getTodoHTML from "./todoHtml";
+
 
 export default function renderProjectTodos(projectId){
     
@@ -7,25 +8,13 @@ export default function renderProjectTodos(projectId){
     const todoContainer = document.querySelector('.todoContainer')
     todoContainer.replaceChildren()
     
-
-   
-    // console.log(projects);
-    // find the project with the matching id
+    
+    // find the project with the matching id and append its todo to page
     const selectedProject = projects.find(element => element.id === projectId)
     
-    // if we cannot find the selected project then render the items from the default project array
-    if(!selectedProject){
-        defaultProjectArr.forEach(todo => {
-            todoContainer.appendChild(todo.getTodoHTML())
-        })
-    }
-    else{
-        selectedProject.arr.forEach(todo => {
-            todoContainer.appendChild(todo.getTodoHTML())
-        });
-    }
-    // console.log(selectedProject.arr);
-    // loop through its internal todo array and call the gettodohtml method
-
+   selectedProject.arr.forEach(todo => {
+        
+        todoContainer.appendChild(getTodoHTML(todo))
+   });
 
 };
