+++ src/template.html
@@ -20,7 +20,7 @@
                 <button class="addProjectBtn" type="submit" value="submit">+</button>
             </form>
             <div class="projectsFolder">
-                <button class="defaultProject projectBtn active">Today</button>
+                <!-- <button class="defaultProject projectBtn active">Today</button> -->
             </div>
         </div>
         <div class="main">
